prevent error on display of git prompt on 10.8
When commits needed to be pushed to a remote repo, an arrow (↑) is displayed in the user's prompt. We run the prompt through some regular expressions to test if it's empty.

Unfortunately, on Mountain Lion, when we test a string that contains the arrow with ZSH's [[ $string =~ 'regex' ]] method, an error is shown on the screen.

I fixed this by using grep -E to test if the string is empty.